finished merging and integration of publication and deliveryArea, along with tests
